+++ src/internal/interpreter/http/security/apiKey/api-key.ts
@@ -13,7 +13,7 @@ import {
   pipeMethod,
   pipeQueryParameters,
   pipeUrl,
-} from '../../http';
+} from '../../pipe';
 import {
   DEFAULT_AUTHORIZATION_HEADER_NAME,
   ISecurityHandler,
